Dockerfile: Use multi-stage docker build to test the image
The buildx setup doesn't work well with `docker run` for multi arch images.
So instead of testing the build after its built, test it during build!
This seems kind of nice anyway as we'll never have a bad build.

The `RUN --mount... /bin/true` hack is needed to force buildx to run the `verify` stage.
If we don't force the `verify` stage, buildx figures out that it doesn't affect the final image
and skips it completely.